Give the modifier test helper classes their own files
Item, ItemWithOrigin and ArrayAccessType were declared at the bottom of PluckTest.php,
so SelectTest only found them if PluckTest.php happened to be loaded into the same
process first. Tests\ is autoloaded from tests/, so a file each is all they need.
